Refactored to be able to use different servers
